MA2/Mobile_and_wireless_networks/esp32_main/Server/requirements.txt
paho-mqtt
pycryptodome
matplotlib
